refactor: denser pool events
